reorganized the previous header nav bar to only show up on the Home component/page. Also moved the logout button from the old header nav bar to only being accessible in the Dashboard component/page
