+++ src/components/Dashboard.js
@@ -1,5 +1,20 @@
 import React from 'react';
+import { Link, useHistory } from 'react-router-dom';
 
 export const Dashboard = () => {
-  return <>Dashboard</>;
+  const history = useHistory();
+
+  const handleLogout = () => {
+    localStorage.clear();
+    history.push('/');
+  };
+
+  return (
+    <>
+      <p>Dashboard</p>
+      <Link to='/' onClick={handleLogout}>
+        Logout
+      </Link>
+    </>
+  );
 };
