+++ src/components/Home.js
@@ -1,5 +1,13 @@
 import React from 'react';
+import { Link } from 'react-router-dom';
 
 export const Home = () => {
-  return <>Home</>;
+  return (
+    <>
+      <nav>
+        <Link to='/login'>Login</Link>
+      </nav>
+      <p>Home</p>
+    </>
+  );
 };
